# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Margie's Barge Report API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [barge]
      summary: Get app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [barge]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /tasks:
    get:
      operationId: listTasks
      tags: [barge]
      summary: List maintenance tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [barge]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [barge]
      summary: Mark a task done
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionCreate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /dock-adjustments:
    post:
      operationId: createDockAdjustment
      tags: [barge]
      summary: Log dock adjustment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DockAdjustmentCreate"
      responses:
        "201":
          description: Dock adjustment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DockAdjustment"
  /bookings:
    get:
      operationId: listBookings
      tags: [barge]
      summary: List lake house bookings
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [barge]
      summary: Create booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreate"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    delete:
      operationId: deleteBooking
      tags: [barge]
      summary: Remove booking
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Removed
  /bring-items:
    get:
      operationId: listBringItems
      tags: [barge]
      summary: List bring-next-time items
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BringItem"
    post:
      operationId: createBringItem
      tags: [barge]
      summary: Add bring item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BringItemCreate"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BringItem"
  /bring-items/{id}:
    delete:
      operationId: deleteBringItem
      tags: [barge]
      summary: Mark bring item brought
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BringItemDelete"
      responses:
        "204":
          description: Removed
  /issues:
    get:
      operationId: listIssues
      tags: [barge]
      summary: List issues
      parameters:
        - name: includeResolved
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [barge]
      summary: Create issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCreate"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}/resolve:
    post:
      operationId: resolveIssue
      tags: [barge]
      summary: Resolve issue
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueResolve"
      responses:
        "200":
          description: Resolved issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Settings:
      type: object
      properties:
        safeLow:
          type: number
        safeHigh:
          type: number
        familyEmails:
          type: array
          items:
            type: string
      required: [safeLow, safeHigh, familyEmails]
    SettingsUpdate:
      type: object
      properties:
        safeLow:
          type: number
        safeHigh:
          type: number
        familyEmails:
          type: array
          items:
            type: string
...
    DockAdjustment:
      type: object
      properties:
        id:
          type: integer
        personName:
          type: string
        workDate:
          type: string
        clearanceUp:
          type: ["number", "null"]
        clearanceDown:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, personName, workDate, clearanceUp, clearanceDown, createdAt]
    DockAdjustmentCreate:
      type: object
      properties:
        personName:
          type: string
        workDate:
          type: string
        clearanceUp:
          type: number
        clearanceDown:
          type: number
      required: [personName, workDate, clearanceUp, clearanceDown]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        icon:
          type: string
        name:
          type: string
        cadenceDays:
          type: integer
        cadenceLabel:
          type: string
        activeMonths:
          type: [string, 'null']
        lastDoneDate:
          type: [string, 'null']
        lastDoneBy:
          type: [string, 'null']
        nextDueDate:
          type: [string, 'null']
        status:
          type: string
          enum: [good, due-soon, overdue, seasonal]
        notes:
          type: [string, 'null']
      required: [id, icon, name, cadenceDays, cadenceLabel, activeMonths, lastDoneDate, lastDoneBy, nextDueDate, status, notes]
    TaskCreate:
      type: object
      properties:
        icon:
          type: string
        name:
          type: string
        cadenceDays:
          type: integer
        activeMonths:
          type: [string, 'null']
      required: [icon, name, cadenceDays]
    CompletionCreate:
      type: object
      properties:
        personName:
          type: string
        workDate:
          type: string
      required: [personName, workDate]
    Booking:
      type: object
      properties:
        id:
          type: integer
        personName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        googleEventId:
          type: [string, 'null']
        createdAt:
          type: string
      required: [id, personName, startDate, endDate, googleEventId, createdAt]
    BookingCreate:
      type: object
      properties:
        personName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        spokeWithUncles:
          type: boolean
        spokeWithCousins:
          type: boolean
      required: [personName, startDate, endDate, spokeWithUncles, spokeWithCousins]
    BringItem:
      type: object
      properties:
        id:
          type: integer
        personName:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required: [id, personName, description, createdAt]
    BringItemCreate:
      type: object
      properties:
        personName:
          type: string
        description:
          type: string
      required: [personName, description]
    BringItemDelete:
      type: object
      properties:
        personName:
          type: string
        broughtDate:
          type: string
      required: [personName, broughtDate]
    Issue:
      type: object
      properties:
        id:
          type: integer
        personName:
          type: string
        caption:
          type: string
        photoUrl:
          type: string
        urgent:
          type: boolean
        status:
          type: string
          enum: [open, resolved]
        resolutionNote:
          type: [string, 'null']
        resolvedBy:
          type: [string, 'null']
        resolvedAt:
          type: [string, 'null']
        createdAt:
          type: string
      required: [id, personName, caption, photoUrl, urgent, status, resolutionNote, resolvedBy, resolvedAt, createdAt]
    IssueCreate:
      type: object
      properties:
        personName:
          type: string
        caption:
          type: string
        photoUrl:
          type: string
        urgent:
          type: boolean
      required: [personName, caption, photoUrl, urgent]
    IssueResolve:
      type: object
      properties:
        personName:
          type: string
        resolutionNote:
          type: string
      required: [personName, resolutionNote]
...