# lib/api-zod/src/generated/types/dockAdjustmentCreate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Margie's Barge Report API
 * OpenAPI spec version: 0.1.0
 */

export interface DockAdjustmentCreate {
  personName: string;
  workDate: string;
  clearanceUp: number;
  clearanceDown: number;
}